feat: Add projectile simulation with PPM image output
This commit adds a projectile simulation that tracks the position of a projectile under the influence of gravity and wind. The projectile's path is plotted on a canvas, which is then written to a PPM image file using a new file-saving function. If no filename is provided to this function, a default filename is generated based on the dimensions of the image and the current date and time.